location patch handler moved to separate module
