+++ server/routes/locations/index.js
@@ -3,12 +3,14 @@ const locationList = require('./locationList');
 const locationGroceries = require('./locationGroceries');
 const locationDelete = require('./locationDelete');
 const locationAdd = require('./locationAdd');
+const locationUpdate = require('./locationUpdate');
 
 const locationRouter = express.Router();
 
 locationRouter.get('/locations', locationList);
 locationRouter.post('/locations', locationAdd);
 locationRouter.get('/locations/:id', locationGroceries);
+locationRouter.patch('/locations/:id', locationUpdate);
 locationRouter.delete('/location/:id', locationDelete);
 
 module.exports = locationRouter;
